Include download/upload file size limitations
I seem to be hitting a file size limit when downloading/uploading blobs, but don't see that limitation documented.  Cloud that be added here?  I don't see an issues area for this repo or I'd have just made an issue for this.  Apologies in advance for submitting a pull request with a blank space.